+++ saleor/tests/fixtures.py
@@ -5665,6 +5665,26 @@ def page(db, page_type):
 
 
 @pytest.fixture
+def two_pages(page):
+    data = {
+        "slug": "test-url-2",
+        "title": "Test page 2",
+        "content": dummy_editorjs("Test content 2."),
+        "is_published": True,
+        "page_type": page.page_type,
+    }
+    page2 = Page.objects.create(**data)
+
+    # associate attribute value to the second page
+    page_attr = page.page_type.page_attributes.first()
+    page_attr_value = page_attr.values.first()
+
+    associate_attribute_values_to_instance(page2, page_attr, page_attr_value)
+
+    return page, page2
+
+
+@pytest.fixture
 def page_with_rich_text_attribute(db, page_type_with_rich_text_attribute):
     data = {
         "slug": "test-url",